# swagger 1104.yaml
openapi: 3.0.3
info:
  title: 4V ODS
  description: |-
    API Specification for 4V ODS
  version: 2.0.0
servers:
  - url: http://localhost:8000
...
paths:
...
  /ods:
    get:
      tags:
        - ODS
      summary: get all ODS
      operationId: getAllODS
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ods"
        "400":
          description: Any problem in the Server
    post:
      tags:
        - ODS
      summary: create new ODS
      operationId: createNewOds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Newods"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Newods"
        "400":
          description: Any problem in the Server
  /ods/{idOds}:
    get:
      tags:
        - ODS
      summary: get ODS by id
      parameters:
        - name: idOds
          in: path
          description: id ODS
          required: true
          schema:
            type: integer
            format: int64

      operationId: getOdsById
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ods"
        "400":
          description: Any problem in the Server
    put:
      tags:
        - ODS
      summary: Update an existing ODS
      operationId: updateOds
      parameters:
        - name: idOds
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateOds"
      responses:
        "200":
          description: Iniciative updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Newods"
        "404":
          description: Iniciative not found
        "400":
          description: Invalid input

    delete:
      tags:
        - ODS
      summary: Delete a ODS
      operationId: deleteOds
      parameters:
        - name: idOds
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Iniciative deleted successfully
        "404":
          description: Iniciative not found
  /goal:
    get:
      tags:
        - Goals
      summary: get all goals
      operationId: getAllGoals
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/goal"
        "400":
          description: Any problem in the Server
    post:
      tags:
        - Goals
      summary: Create new goal
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newGoal"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/goal"
        "400":
          description: Any problem in the Server
  /goal/{idGoal}:
    get:
      tags:
        - Goals
      summary: get goal by id
      parameters:
        - name: idGoal
          in: path
          description: id goal
          required: true
          schema:
            type: integer
            format: int64

      operationId: getGoalById
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/goal"
        "400":
          description: Any problem in the Server
    put:
      tags:
        - Goals
      summary: Update an existing goal
      operationId: updateGoal
      parameters:
        - name: idGoal
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newGoal"
      responses:
        "200":
          description: Iniciative updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/newGoal"
        "404":
          description: Iniciative not found
        "400":
          description: Invalid input

    delete:
      tags:
        - Goals
      summary: Delete a goal
      operationId: deleteGoal
      parameters:
        - name: idGoal
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Iniciative deleted successfully
        "404":
          description: Iniciative not found
  /company:
    get:
      tags:
        - Company
      summary: get all companies
      operationId: getAllCompanies
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/company"
        "400":
          description: Any problem in the Server
    post:
      tags:
        - Company
      summary: create Company
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newCompany"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/company"

        "400":
          description: Any problem in the Server
  /company/{idCompany}:
    get:
      tags:
        - Company
      summary: get company by id
      parameters:
        - name: idCompany
          in: path
          description: id company
          required: true
          schema:
            type: integer
            format: int64

      operationId: getCompanyById
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/company"
        "400":
          description: Any problem in the Server
    put:
      tags:
        - Company
      summary: update a Company
      operationId: updateCompany
      parameters:
        - name: idCompany
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/company"

        "400":
          description: Any problem in the Server
  /company/{id}:
    delete:
      tags:
        - Company
      summary: Delete a company
      operationId: deleteCompany
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Iniciative deleted successfully
        "404":
          description: Iniciative not found
  /teacher:
    get:
      tags:
        - Teacher
      summary: get all teachers
      operationId: getAllTeachers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
        "400":
          description: Any problem in the Server
    post:
      tags:
        - Teacher
      summary: create Teacher
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newTeacher"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"

        "400":
          description: Any problem in the Server
  /teacher/{idTeacher}:
    get:
      tags:
        - Teacher
      summary: get teacher by id
      parameters:
        - name: idTeacher
          in: path
          description: id teacher
          required: true
          schema:
            type: integer
            format: int64

      operationId: getTeacherById
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
        "400":
          description: Any problem in the Server
    put:
      tags:
        - Teacher
      summary: update a Teacher
      operationId: updateTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newTeacher"
      parameters:
        - name: idTeacher
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"

        "400":
          description: Any problem in the Server
    delete:
      tags:
        - Teacher
      summary: Delete a teacher
      operationId: deleteTeacher
      parameters:
        - name: idTeacher
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Iniciative deleted successfully
        "404":
          description: Iniciative not found
...
components:
  schemas:
...
    ods:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        description:
          type: string
          example: Hambre en el mundo
        dimension:
          type: string
          example: Social
    Newods:
      type: object
      properties:
        description:
          type: string
          example: ODS 5
        goals:
          type: array
          items:
            type: string
            example: Nueva meta a incorporar

        dimension:
          type: string
          example: Social
    updateOds:
      type: object
      properties:
        description:
          type: string
          example: ODS 5

        dimension:
          type: string
          example: Social
    goal:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        description:
          type: string
          example: Reducir la contaminación un 10%
        ods:
          type: integer
          format: int64
          example: 2
    newGoal:
      type: object
      properties:
        description:
          type: string
          example: Reducir la contaminación un 10%
        idOds:
          type: integer
          format: int64
          example: 2
    company:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Green Energy CO
    newCompany:
      type: object
      properties:
        name:
          type: string
          example: Green Energy CO
    teacher:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Carlos Gutierrez
    newTeacher:
      type: object
      properties:
        name:
          type: string
          example: Carlos Gutierrez
...